+++ src/features/Statistics.jsx
@@ -1,4 +1,9 @@
+import { useSelector } from 'react-redux';
+import { selectClickedClues } from './categorySlice';
+
 const Statistics = () => {
+  const answeredQuestions = useSelector(selectClickedClues);
+  console.log(answeredQuestions);
   return <div>Statistics</div>;
 };
